fix:product 조회 스칼라 서브 쿼리로 수정

+++ src/main/java/com/danggeun/market/interest/service/InterestProductSearchService.java
@@ -30,7 +30,7 @@ public class InterestProductSearchService {
         List<Object[]> result = interestHistoryRepository.searchInterestProduct(user);
         return result.stream()
                 .map(objects -> new ProductSummaryResponse(((InterestHistory) objects[0]).getProduct(),
-                        (Long) objects[1],
+                        (Integer) objects[1],
                         (Long) objects[2]))
                 .collect(Collectors.toList());
     }
